+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
+
 Changelog to track updates for this mod.  
-    Add your changes to Unreleased if you want to commit.  
-    Please write according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
+Add your changes to Unreleased if you want to commit.  
+Please write according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
 ### Added
+
+- Add parabolic antenna craft recipe
 
 ### Changed
 
@@ -20,13 +23,17 @@ Changelog to track updates for this mod.
 ## [3.19.6] - 2023-01-21
 
 ### Fixed
+
 - Fixed a bug that could occur when reloading while playing music
 
 ## [3.19.5] - 2023-01-17
 
 ### Fixed
+
 - Fixed server crash due to `NullPointerException` [Issues#45](https://github.com/TeamFelnull/IamMusicPlayer/issues/45)
 
 [Unreleased]: https://github.com/TeamFelnull/IamMusicPlayer/compare/v3.19.6...HEAD
+
 [3.19.6]: https://github.com/TeamFelnull/IamMusicPlayer/compare/v3.19.5...v3.19.6
+
 [3.19.5]: https://github.com/TeamFelnull/IamMusicPlayer/commits/v3.19.5